refactor(logic): move safe gas meter to meter pkg
