+++ x/logic/meter/safe.go
@@ -1,4 +1,4 @@
-package types
+package meter
 
 import (
 	"runtime"
@@ -7,17 +7,23 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-var _ sdk.GasMeter = (*safeGasMeter)(nil)
-
-// safeGasMeter is a wrapper around sdk.GasMeter that provides go-routine-safe access to the underlying gas meter.
+// safeMeterDecorater is a wrapper around sdk.GasMeter that provides go-routine-safe access to the underlying gas meter.
 // This is needed because the interpreter is uses multiple go-routines, and the gas meter is shared between multiple
 // goroutines.
-type safeGasMeter struct {
+type safeMeterDecorater struct {
 	gasMeter sdk.GasMeter
 	mutex    sync.RWMutex
 }
 
-func (m *safeGasMeter) ConsumeGas(amount uint64, descriptor string) {
+// WithSafeMeter returns a new instance of sdk.GasMeter that is go-routine-safe.
+func WithSafeMeter(gasMeter sdk.GasMeter) sdk.GasMeter {
+	return &safeMeterDecorater{
+		gasMeter: gasMeter,
+	}
+}
+
+// ConsumeGas consumes the given amount of gas from the decorated gas meter.
+func (m *safeMeterDecorater) ConsumeGas(amount uint64, descriptor string) {
 	m.mutex.Lock()
 	defer func() {
 		if r := recover(); r != nil {
@@ -38,65 +44,66 @@ func (m *safeGasMeter) ConsumeGas(amount uint64, descriptor string) {
 	m.gasMeter.ConsumeGas(amount, descriptor)
 }
 
-func (m *safeGasMeter) GasConsumed() uint64 {
+// GasConsumed returns the amount of gas consumed by the decorated gas meter.
+func (m *safeMeterDecorater) GasConsumed() uint64 {
 	m.mutex.RLocker().Lock()
 	defer m.mutex.RLocker().Unlock()
 
 	return m.gasMeter.GasConsumed()
 }
 
-func (m *safeGasMeter) GasConsumedToLimit() uint64 {
+// GasConsumedToLimit returns the amount of gas consumed by the decorated gas meter.
+func (m *safeMeterDecorater) GasConsumedToLimit() uint64 {
 	m.mutex.RLocker().Lock()
 	defer m.mutex.RLocker().Unlock()
 
 	return m.gasMeter.GasConsumedToLimit()
 }
 
-func (m *safeGasMeter) IsPastLimit() bool {
+// IsPastLimit returns true if the gas limit has been reached.
+func (m *safeMeterDecorater) IsPastLimit() bool {
 	m.mutex.RLocker().Lock()
 	defer m.mutex.RLocker().Unlock()
 
 	return m.gasMeter.IsPastLimit()
 }
 
-func (m *safeGasMeter) Limit() uint64 {
+// Limit returns the gas limit of the decorated gas meter.
+func (m *safeMeterDecorater) Limit() uint64 {
 	m.mutex.RLocker().Lock()
 	defer m.mutex.RLocker().Unlock()
 
 	return m.gasMeter.Limit()
 }
 
-func (m *safeGasMeter) RefundGas(amount uint64, descriptor string) {
+// RefundGas refunds the given amount of gas to the decorated gas meter.
+func (m *safeMeterDecorater) RefundGas(amount uint64, descriptor string) {
 	m.mutex.Lock()
 	defer m.mutex.Unlock()
 
 	m.gasMeter.RefundGas(amount, descriptor)
 }
 
-func (m *safeGasMeter) GasRemaining() uint64 {
+// GasRemaining returns the amount of gas remaining in the decorated gas meter.
+func (m *safeMeterDecorater) GasRemaining() uint64 {
 	m.mutex.RLocker().Lock()
 	defer m.mutex.RLocker().Unlock()
 
 	return m.gasMeter.GasRemaining()
 }
 
-func (m *safeGasMeter) String() string {
+// String returns a string representation of the decorated gas meter.
+func (m *safeMeterDecorater) String() string {
 	m.mutex.RLocker().Lock()
 	defer m.mutex.RLocker().Unlock()
 
 	return m.gasMeter.String()
 }
 
-func (m *safeGasMeter) IsOutOfGas() bool {
+// IsOutOfGas returns true if the gas limit has been reached.
+func (m *safeMeterDecorater) IsOutOfGas() bool {
 	m.mutex.RLocker().Lock()
 	defer m.mutex.RLocker().Unlock()
 
 	return m.gasMeter.IsOutOfGas()
-}
-
-// NewSafeGasMeter returns a new instance of sdk.GasMeter that is go-routine-safe.
-func NewSafeGasMeter(gasMeter sdk.GasMeter) sdk.GasMeter {
-	return &safeGasMeter{
-		gasMeter: gasMeter,
-	}
 }
